Fix Chip padding issue
