+++ lib/src/views/setup/_1_tournament_name.dart
@@ -20,7 +20,7 @@ import 'package:friends_tournament/src/provider/setup_provider.dart';
 import 'package:friends_tournament/src/style/app_style.dart';
 import 'package:friends_tournament/src/ui/text_field_decoration.dart';
 import 'package:friends_tournament/src/utils/app_localizations.dart';
-import 'package:friends_tournament/src/views/setup/setup_chip_separator.dart';
+import 'package:friends_tournament/src/ui/chip_separator.dart';
 import 'package:friends_tournament/src/views/setup/setup_page.dart';
 import 'package:provider/provider.dart';
 
@@ -44,7 +44,7 @@ class TournamentName extends StatelessWidget implements SetupPage {
                   child: _buildImage(),
                 ),
                 _buildTitle(context),
-                const SetupChipSeparator(),
+                _buildChipSeparator(),
                 _buildTextField(context),
                 Expanded(
                   flex: 4,
@@ -75,6 +75,16 @@ class TournamentName extends StatelessWidget implements SetupPage {
         AppLocalizations.translate(context, 'tournament_name_title'),
         style: AppTextStyle.onboardingTitleStyle,
       ),
+    );
+  }
+
+  Widget _buildChipSeparator() {
+    return const Padding(
+      padding: EdgeInsets.only(
+        top: MarginsRaw.medium,
+        bottom: MarginsRaw.medium,
+      ),
+      child: ChipSeparator(),
     );
   }
