+++ lib/src/views/setup/_3_player_ast_number.dart
@@ -20,7 +20,7 @@ import 'package:friends_tournament/src/provider/setup_provider.dart';
 import 'package:friends_tournament/src/style/app_style.dart';
 import 'package:friends_tournament/src/ui/setup_counter_widget.dart';
 import 'package:friends_tournament/src/utils/app_localizations.dart';
-import 'package:friends_tournament/src/views/setup/setup_chip_separator.dart';
+import 'package:friends_tournament/src/ui/chip_separator.dart';
 import 'package:friends_tournament/src/views/setup/setup_page.dart';
 import 'package:provider/provider.dart';
 
@@ -42,7 +42,7 @@ class PlayersAST extends StatelessWidget implements SetupPage {
                 child: _buildImage(),
               ),
               _buildTitle(context),
-              const SetupChipSeparator(),
+              _buildChipSeparator(),
               _buildPlayerASTCounter(),
               Expanded(
                 flex: 4,
@@ -72,6 +72,16 @@ class PlayersAST extends StatelessWidget implements SetupPage {
         AppLocalizations.translate(context, 'number_of_players_ast_title'),
         style: AppTextStyle.onboardingTitleStyle,
       ),
+    );
+  }
+
+  Widget _buildChipSeparator() {
+    return const Padding(
+      padding: EdgeInsets.only(
+        top: MarginsRaw.medium,
+        bottom: MarginsRaw.medium,
+      ),
+      child: ChipSeparator(),
     );
   }
