+++ lib/src/views/setup/_4_matches_number.dart
@@ -21,7 +21,7 @@ import 'package:friends_tournament/src/provider/setup_provider.dart';
 import 'package:friends_tournament/src/style/app_style.dart';
 import 'package:friends_tournament/src/ui/setup_counter_widget.dart';
 import 'package:friends_tournament/src/utils/app_localizations.dart';
-import 'package:friends_tournament/src/views/setup/setup_chip_separator.dart';
+import 'package:friends_tournament/src/ui/chip_separator.dart';
 import 'package:friends_tournament/src/views/setup/setup_page.dart';
 import 'package:provider/provider.dart';
 
@@ -43,7 +43,7 @@ class MatchesNumber extends StatelessWidget implements SetupPage {
                   child: _buildImage(),
                 ),
                 _buildTitle(context),
-                const SetupChipSeparator(),
+                _buildChipSeparator(),
                 _buildMatchNumberCounter(),
                 Expanded(
                   flex: 4,
@@ -74,6 +74,16 @@ class MatchesNumber extends StatelessWidget implements SetupPage {
         AppLocalizations.translate(context, 'number_of_matches_title'),
         style: AppTextStyle.onboardingTitleStyle,
       ),
+    );
+  }
+
+  Widget _buildChipSeparator() {
+    return const Padding(
+      padding: EdgeInsets.only(
+        top: MarginsRaw.medium,
+        bottom: MarginsRaw.medium,
+      ),
+      child: ChipSeparator(),
     );
   }
