+++ lib/src/views/setup/_6_matches_name.dart
@@ -24,7 +24,7 @@ import 'package:friends_tournament/src/style/app_style.dart';
 import 'package:friends_tournament/src/ui/text_field_tile.dart';
 import 'package:friends_tournament/src/utils/app_localizations.dart';
 import 'package:friends_tournament/src/utils/widget_keys.dart';
-import 'package:friends_tournament/src/views/setup/setup_chip_separator.dart';
+import 'package:friends_tournament/src/ui/chip_separator.dart';
 import 'package:friends_tournament/src/views/setup/setup_page.dart';
 import 'package:provider/provider.dart';
 
@@ -99,7 +99,7 @@ class MatchesName extends StatelessWidget implements SetupPage {
           child: _buildImage(),
         ),
         _buildTitle(context),
-        const SetupChipSeparator(),
+        _buildChipSeparator(),
         Expanded(
           flex: 7,
           child: _buildTextFields(),
@@ -127,6 +127,16 @@ class MatchesName extends StatelessWidget implements SetupPage {
         AppLocalizations.translate(context, 'matches_name_title'),
         style: AppTextStyle.onboardingTitleStyle,
       ),
+    );
+  }
+
+  Widget _buildChipSeparator() {
+    return const Padding(
+      padding: EdgeInsets.only(
+          top: MarginsRaw.regular,
+          left: MarginsRaw.regular,
+          right: MarginsRaw.regular),
+      child: ChipSeparator(),
     );
   }
