+++ lib/src/views/tournament/final_screen.dart
@@ -20,6 +20,7 @@ import 'package:flutter_svg/svg.dart';
 import 'package:friends_tournament/src/provider/leaderboard_provider.dart';
 import 'package:friends_tournament/src/provider/setup_provider.dart';
 import 'package:friends_tournament/src/style/app_style.dart';
+import 'package:friends_tournament/src/ui/chip_separator.dart';
 import 'package:friends_tournament/src/ui/error_dialog.dart';
 import 'package:friends_tournament/src/utils/app_localizations.dart';
 import 'package:friends_tournament/src/utils/widget_keys.dart';
@@ -147,23 +148,13 @@ class _FinalScreenState extends State<FinalScreen> {
   }
 
   Widget _buildChipSeparator() {
-    return Padding(
-      padding: const EdgeInsets.only(
+    return const Padding(
+      padding: EdgeInsets.only(
         top: MarginsRaw.regular,
         left: MarginsRaw.regular,
         bottom: MarginsRaw.medium,
       ),
-      child: Container(
-        alignment: Alignment.topLeft,
-        decoration: BoxDecoration(
-          color: AppColors.blue,
-          borderRadius: BorderRadius.circular(
-            MarginsRaw.borderRadius,
-          ),
-        ),
-        height: 6,
-        width: 60,
-      ),
+      child: ChipSeparator(),
     );
   }
