+++ lib/src/views/welcome_screen.dart
@@ -19,6 +19,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:friends_tournament/src/provider/setup_provider.dart';
 import 'package:friends_tournament/src/style/app_style.dart';
+import 'package:friends_tournament/src/ui/chip_separator.dart';
 import 'package:friends_tournament/src/utils/app_localizations.dart';
 import 'package:friends_tournament/src/views/settings/settings_screen.dart';
 import 'package:friends_tournament/src/views/setup/setup_pages_container.dart';
@@ -139,19 +140,11 @@ class Welcome extends StatelessWidget {
   }
 
   Widget _buildChipSeparator() {
-    return Padding(
-      padding: const EdgeInsets.only(
+    return const Padding(
+      padding: EdgeInsets.only(
         top: MarginsRaw.regular,
       ),
-      child: Container(
-        alignment: Alignment.topLeft,
-        decoration: BoxDecoration(
-          color: AppColors.blue,
-          borderRadius: BorderRadius.circular(MarginsRaw.borderRadius),
-        ),
-        height: 6,
-        width: 60,
-      ),
+      child: ChipSeparator(),
     );
   }
